fixed out-of-order definition problem
symbols were previously parsed in such a way that writing the symbols in declaration order caused a definition problem.  this parser now unpermutes the cell arrays for constants, functions, parameters, and subfunctions such that they are ordered by definition.  thus, they can just be written later in that order, from top to bottom.

this problem was fixed due to an out-of-order issue in a 4-1 alt-burmester input file being tested for the bertini_real-chebfun interface.  yay.